fix(PDP.fix): fix fetch, add custom hook fo Product Page
