+++ src/pages/FavouritesPage/FavouritesPage.module.scss
@@ -8,10 +8,10 @@
     justify-content: center;
     align-items: center;
     font-weight: 800;
-    font-size: 32px;
-    line-height: 30px;
-    height: 80vh;
-    color: rgb(255, 255, 255);
+    font-size: 24px;
+    line-height: 24px;
+    color: var(--main-second);
+    height: 50vh;
   }
 
   .breadcrumbs {
